Keyboard + UI Colors Updated
The Keyboard does not use the Button Component, but rather relies on script to enhance the player exprerience. It is now possible to play notes fast one after another. Besides that, the notes play as long as you can hold them, instead of playing for a fixed amount of time.